+++ CHANGES.md
@@ -4,6 +4,22 @@
 
 - **Breaking:** Update to GAP 4.14.0
 
+- **Breaking:** Change the default behaviour of `gap_to_julia`
+  w.r.t. the recursive conversion of subobjects.
+  Now `gap_to_julia(T, obj)` converts subobjects of `obj` only until the
+  result has the type `T`.
+  In order to convert subobjects of `obj` recursively,
+  one can call `gap_to_julia(T, obj; recursive = true)`.
+  In earlier versions, recursive conversion was the default,
+  and `gap_to_julia(T, obj; recursive = false)` was used to request
+  non-recursive conversion.
+
+- **Breaking:** Do not support the optional `GapCacheDict` argument for
+  `GapObj` anymore, because recursive conversion to GAP is handled by
+  `GAP.GapObj_internal` and does not involve cals to `GapObj`.
+  We had always stated in the documentation that users should not enter
+  this argument because it gets created automatically in recursive conversions.
+
 ## Version 0.15.3 (released 2025-09-24)
 
 - Drop dependency on `Pkg`.